requirements.txt
dnspython3
requests
urllib3
